refactor: 리뷰 반영 코드 및 파일 수정 - map-service를 parking-service에 통합 - ParkingLotSimpleinfoModal.tsx 에서 fetch훅을 사용하여 데이터를 불러오도록 변경 - RatingOverview에 값이 넘어오지 않는 경우에는 표시하지 않도록 변경 - 미사용 및 중복 type 선언 및 dummy data 제거 #122
