+++ apps/client/src/components/pages/map/RatingOverview.tsx
@@ -1,10 +1,10 @@
 import { Star, ThumbsDown, ThumbsUp } from 'lucide-react';
 
 export default function RatingOverview({
-  averageRating = 0,
-  reviewCount = 0,
-  likeCount = 0,
-  dislikeCount = 0,
+  averageRating,
+  reviewCount,
+  likeCount,
+  dislikeCount,
 }: {
   averageRating?: number;
   reviewCount?: number;
@@ -13,19 +13,27 @@ export default function RatingOverview({
 }) {
   return (
     <div className="flex space-x-2 text-13px">
-      <p className="flex items-center gap-1">
-        <Star fill="currentColor" className="text-yellow-2" size={14} />
-        {averageRating}
-        {reviewCount < 999 ? `(${reviewCount})` : `(999+)`}
-      </p>
-      <p className="flex items-center gap-1">
-        <ThumbsUp fill="currentColor" className="text-gray-2" size={14} />
-        {likeCount < 999 ? likeCount : `999+`}
-      </p>
-      <p className="flex items-center gap-1">
-        <ThumbsDown fill="currentColor" className="text-gray-2" size={14} />
-        {dislikeCount < 999 ? dislikeCount : `999+`}
-      </p>
+      {(averageRating !== undefined || reviewCount !== undefined) && (
+        <p className="flex items-center gap-0.5">
+          <Star fill="currentColor" className="text-yellow-2" size={14} />
+          {typeof averageRating === 'number' && <span>{averageRating}</span>}
+          {typeof reviewCount === 'number' && (
+            <span>{reviewCount < 999 ? `(${reviewCount})` : `(999+)`}</span>
+          )}
+        </p>
+      )}
+      {likeCount !== undefined && (
+        <p className="flex items-center gap-0.5">
+          <ThumbsUp fill="currentColor" className="text-gray-2" size={14} />
+          {likeCount < 999 ? likeCount : `999+`}
+        </p>
+      )}
+      {dislikeCount !== undefined && (
+        <p className="flex items-center gap-0.5">
+          <ThumbsDown fill="currentColor" className="text-gray-2" size={14} />
+          {dislikeCount < 999 ? dislikeCount : `999+`}
+        </p>
+      )}
     </div>
   );
 }
