--- apps/client/src/data/markerDummyData.ts
@@ -1,322 +1,0 @@
-import { MarkerDataType, ParkingLotModalInfoType } from '@/types/mapDataTypes';
-
-const rawPositions = [
-  { lat: 35.14538871, lng: 129.19709887 },
-  { lat: 35.11675112, lng: 128.94440006 },
-  { lat: 35.0891994, lng: 128.95189525 },
-  { lat: 35.15433391, lng: 129.11662195 },
-  { lat: 35.09229203, lng: 128.93523497 },
-  { lat: 35.13801613, lng: 129.17611338 },
-  { lat: 35.20082535, lng: 129.06749988 },
-  { lat: 35.18380178, lng: 129.13415404 },
-  { lat: 35.23417469, lng: 129.07858205 },
-  { lat: 35.1995266, lng: 129.1595313 },
-  { lat: 35.18268486, lng: 129.04960386 },
-  { lat: 35.11243472, lng: 129.18686109 },
-  { lat: 35.06181374, lng: 129.1595332 },
-  { lat: 35.14444924, lng: 128.93230941 },
-  { lat: 35.2469464, lng: 128.96588422 },
-  { lat: 35.23708724, lng: 128.92224613 },
-  { lat: 35.2048481, lng: 128.94790357 },
-  { lat: 35.21319473, lng: 129.16107488 },
-  { lat: 35.12837346, lng: 129.04963007 },
-  { lat: 35.19685704, lng: 129.0426815 },
-  { lat: 35.05523023, lng: 129.19123122 },
-  { lat: 35.0625386, lng: 128.90767867 },
-  { lat: 35.05084405, lng: 129.0004224 },
-  { lat: 35.19321971, lng: 129.08905587 },
-  { lat: 35.2171948, lng: 128.91208388 },
-  { lat: 35.09338468, lng: 128.98401519 },
-  { lat: 35.12304149, lng: 129.12552362 },
-  { lat: 35.1029751, lng: 128.93582047 },
-  { lat: 35.20344779, lng: 129.01924885 },
-  { lat: 35.19075481, lng: 128.98867326 },
-  { lat: 35.09780906, lng: 129.10539896 },
-  { lat: 35.17326406, lng: 128.9387984 },
-  { lat: 35.14215273, lng: 129.04041077 },
-  { lat: 35.05783128, lng: 128.96534444 },
-  { lat: 35.2465227, lng: 129.19083712 },
-  { lat: 35.0971898, lng: 128.96833406 },
-  { lat: 35.10551663, lng: 129.03023187 },
-  { lat: 35.22784019, lng: 128.97407269 },
-  { lat: 35.15604905, lng: 129.11844183 },
-  { lat: 35.08498571, lng: 128.95943193 },
-  { lat: 35.15336343, lng: 129.11300075 },
-  { lat: 35.15239735, lng: 129.0510761 },
-  { lat: 35.21602573, lng: 129.03864889 },
-  { lat: 35.12438643, lng: 129.14081949 },
-  { lat: 35.10661716, lng: 128.94119651 },
-  { lat: 35.14098282, lng: 128.97342155 },
-  { lat: 35.23684271, lng: 128.97864469 },
-  { lat: 35.16685939, lng: 128.97165387 },
-  { lat: 35.20745562, lng: 129.0534016 },
-  { lat: 35.20326269, lng: 129.0125568 },
-  { lat: 35.07435592, lng: 128.92039916 },
-  { lat: 35.14556896, lng: 129.11245929 },
-  { lat: 35.21904632, lng: 128.94820144 },
-  { lat: 35.19230766, lng: 128.9697066 },
-  { lat: 35.11029244, lng: 129.15754026 },
-  { lat: 35.05495773, lng: 129.17972647 },
-  { lat: 35.2410544, lng: 129.1390584 },
-  { lat: 35.09433893, lng: 128.93288781 },
-  { lat: 35.20683975, lng: 129.17202525 },
-  { lat: 35.20915713, lng: 128.94345287 },
-  { lat: 35.12467656, lng: 128.9413337 },
-  { lat: 35.1681994, lng: 128.99220959 },
-  { lat: 35.14251956, lng: 128.9573136 },
-  { lat: 35.14088324, lng: 129.02012889 },
-  { lat: 35.17316252, lng: 129.12276574 },
-  { lat: 35.24587106, lng: 128.90884177 },
-  { lat: 35.08235478, lng: 129.14486168 },
-  { lat: 35.129782, lng: 129.19818732 },
-  { lat: 35.11913306, lng: 129.01800293 },
-  { lat: 35.05959865, lng: 129.02379892 },
-  { lat: 35.12372528, lng: 128.97299952 },
-  { lat: 35.23179375, lng: 129.16013422 },
-  { lat: 35.09680197, lng: 128.91007615 },
-  { lat: 35.07030628, lng: 129.00463316 },
-  { lat: 35.16210807, lng: 129.04912529 },
-  { lat: 35.06630523, lng: 129.1681733 },
-  { lat: 35.14570083, lng: 129.11584352 },
-  { lat: 35.0838654, lng: 129.05060613 },
-  { lat: 35.17839711, lng: 129.19416623 },
-  { lat: 35.20020756, lng: 129.03909953 },
-  { lat: 35.09452381, lng: 129.03053127 },
-  { lat: 35.23490498, lng: 128.92084759 },
-  { lat: 35.06894438, lng: 129.00742839 },
-  { lat: 35.23438401, lng: 129.05139472 },
-  { lat: 35.11043451, lng: 129.13587876 },
-  { lat: 35.15056966, lng: 129.04099506 },
-  { lat: 35.06472821, lng: 128.94460653 },
-  { lat: 35.23855446, lng: 129.03252856 },
-  { lat: 35.20479998, lng: 129.02378252 },
-  { lat: 35.20983087, lng: 128.94339367 },
-  { lat: 35.0692397, lng: 129.16947496 },
-  { lat: 35.18407098, lng: 129.16536412 },
-  { lat: 35.17944775, lng: 129.03710256 },
-  { lat: 35.20107936, lng: 128.9741907 },
-  { lat: 35.14183061, lng: 129.10040896 },
-  { lat: 35.09566461, lng: 129.10937051 },
-  { lat: 35.17910708, lng: 129.16022948 },
-  { lat: 35.23177962, lng: 128.96679236 },
-  { lat: 35.2202767, lng: 128.93192118 },
-  { lat: 35.21279426, lng: 128.97392348 },
-  { lat: 38.83455065787558, lng: 126.40843685209701 },
-  { lat: 39.057590811785516, lng: 129.96830797518857 },
-  { lat: 37.32860533720713, lng: 128.21190659393764 },
-  { lat: 39.48295530948809, lng: 129.55555901757637 },
-  { lat: 35.7702481579169, lng: 127.82459343958025 },
-  { lat: 39.75553943725134, lng: 127.20833722821251 },
-  { lat: 39.88568621508284, lng: 128.67157063591094 },
-  { lat: 37.73307876960214, lng: 128.63506383981357 },
-  { lat: 37.962615073182235, lng: 127.29118271063088 },
-  { lat: 36.05141931338508, lng: 127.1241864676971 },
-  { lat: 36.2202803392731, lng: 129.53540921584406 },
-  { lat: 37.050256598110764, lng: 128.03933946117192 },
-  { lat: 35.87873511497785, lng: 127.5714431378011 },
-  { lat: 36.17308428556241, lng: 127.95931580240497 },
-  { lat: 35.362660698835754, lng: 126.11499524420934 },
-  { lat: 39.403268399811104, lng: 127.1845310231846 },
-  { lat: 38.846265430598464, lng: 127.57180290679904 },
-  { lat: 35.6163646759954, lng: 129.94440860730637 },
-  { lat: 38.00397906438033, lng: 126.54982049835672 },
-  { lat: 36.000238801931026, lng: 126.11436807133877 },
-  { lat: 36.23996375622548, lng: 129.21423311072607 },
-  { lat: 36.85099367711497, lng: 129.14079088338306 },
-  { lat: 38.64905911882487, lng: 127.9835104227717 },
-  { lat: 36.649917602312584, lng: 128.55177735028704 },
-  { lat: 35.64632745555914, lng: 127.41797634469049 },
-  { lat: 39.48168512805678, lng: 128.80068530781855 },
-  { lat: 39.73452099225666, lng: 127.68629293631324 },
-  { lat: 39.592109965107504, lng: 126.47757908384413 },
-  { lat: 35.95200568662633, lng: 128.84761587594468 },
-  { lat: 38.21583772666757, lng: 127.28560411489657 },
-  { lat: 37.14078188309783, lng: 128.3200883913512 },
-  { lat: 38.3348225252669, lng: 127.60025583605679 },
-  { lat: 35.108467606986714, lng: 129.75553817270648 },
-  { lat: 35.32776032188684, lng: 126.56082900332497 },
-  { lat: 36.667806172346985, lng: 128.25940965023318 },
-  { lat: 36.10519625722552, lng: 126.49698295391872 },
-  { lat: 38.41494052514251, lng: 128.03955896272166 },
-  { lat: 36.662977617659614, lng: 126.30540309999002 },
-  { lat: 39.051912562943066, lng: 129.9851246686157 },
-  { lat: 37.82001781720546, lng: 128.71938214702783 },
-  { lat: 39.14111338761312, lng: 126.63943163120159 },
-  { lat: 37.38490469747991, lng: 127.49488645305904 },
-  { lat: 39.07870508041358, lng: 126.20623812060451 },
-  { lat: 39.21261079328445, lng: 126.8235071137053 },
-  { lat: 35.443839745331175, lng: 129.8581532643604 },
-  { lat: 35.84522481049793, lng: 129.82642204456224 },
-  { lat: 35.573835388901614, lng: 128.6418495330732 },
-  { lat: 35.6574486759165, lng: 126.21283391505217 },
-  { lat: 38.58182653279657, lng: 129.7992229998672 },
-  { lat: 39.33851597415878, lng: 128.00179729251474 },
-  { lat: 35.23816711962087, lng: 126.9218025387089 },
-  { lat: 39.956364068231046, lng: 126.51424085143107 },
-  { lat: 37.676414651333786, lng: 127.48299234247163 },
-  { lat: 37.57283571623754, lng: 129.1335487545069 },
-  { lat: 36.78240646132881, lng: 128.0082619428281 },
-  { lat: 37.5419596567968, lng: 127.69882195736388 },
-  { lat: 37.6749541438705, lng: 126.16518914290012 },
-  { lat: 38.44728486726594, lng: 128.8712186275568 },
-  { lat: 39.82813716031499, lng: 126.04770202074779 },
-  { lat: 37.03864165866965, lng: 128.3936700754257 },
-  { lat: 35.327863736498095, lng: 128.0641022940925 },
-  { lat: 36.51571478893083, lng: 126.70771426255718 },
-  { lat: 36.135401102513256, lng: 127.47586331176494 },
-  { lat: 36.93665057985413, lng: 126.41307549724266 },
-  { lat: 37.62781600020715, lng: 129.62144830336695 },
-  { lat: 35.03211993905668, lng: 126.25477241330405 },
-  { lat: 38.2920968032533, lng: 128.48443185617262 },
-  { lat: 38.343361569721054, lng: 129.20396217923488 },
-  { lat: 35.43443585932569, lng: 126.31004969688472 },
-  { lat: 38.078368445201406, lng: 127.18320460918936 },
-  { lat: 38.57764809077302, lng: 127.06901555234715 },
-  { lat: 36.21548117792394, lng: 126.67867022397189 },
-  { lat: 35.69875731701775, lng: 128.83787265251848 },
-  { lat: 36.744189666627655, lng: 126.4339396941532 },
-  { lat: 36.61936397467331, lng: 127.18347835372056 },
-  { lat: 36.166534265491926, lng: 129.35048702962123 },
-  { lat: 35.174439703647636, lng: 129.0823466953374 },
-  { lat: 37.00342500218268, lng: 128.07402077914415 },
-  { lat: 37.294780616545424, lng: 129.3177600426012 },
-  { lat: 35.775802702098, lng: 128.57716052789453 },
-  { lat: 38.48080200927931, lng: 127.77984312885809 },
-  { lat: 37.13708091735384, lng: 128.10937612426503 },
-  { lat: 38.61882945280737, lng: 127.99687902524492 },
-  { lat: 36.53144967607249, lng: 128.90254165606598 },
-  { lat: 37.61021550648583, lng: 128.75853047535418 },
-  { lat: 36.441592283005285, lng: 128.43461346953424 },
-  { lat: 35.516613063544014, lng: 127.39714970599846 },
-  { lat: 38.513253329438896, lng: 128.6789745387321 },
-  { lat: 37.88258553452197, lng: 129.04097371618192 },
-  { lat: 39.18871550426617, lng: 128.33513973448316 },
-  { lat: 38.19031847985528, lng: 128.15513997018232 },
-  { lat: 38.32666160986006, lng: 127.99239726338487 },
-  { lat: 36.1158866512481, lng: 128.96903286302984 },
-  { lat: 35.859256471556364, lng: 128.5519050184046 },
-  { lat: 37.00103190254704, lng: 127.17045641254863 },
-  { lat: 38.4981707807572, lng: 129.7820376522566 },
-  { lat: 37.5591566, lng: 126.9253652 },
-  { lat: 35.1385425, lng: 129.1014781 },
-  { lat: 35.20618517638034, lng: 129.07944301057026 },
-  { lat: 37.561110808242056, lng: 126.9831268386891 },
-  { lat: 37.86187129655063, lng: 127.7410250820423 },
-  { lat: 35.10233410927457, lng: 129.02611815856181 },
-  { lat: 36.33884892276137, lng: 127.393666019664 },
-  { lat: 37.520412849636, lng: 126.9742764161581 },
-  { lat: 35.155139675209675, lng: 129.06154773758374 },
-  { lat: 35.816041994696576, lng: 127.11046706211324 },
-  { lat: 38.20441110638504, lng: 128.59038671285234 },
-  { lat: 37.586112739308916, lng: 127.02949148517999 },
-  { lat: 36.362321615174835, lng: 127.35000483225389 },
-  { lat: 37.55227862908755, lng: 126.92280546294998 },
-  { lat: 37.490413948014606, lng: 127.02079678472444 },
-  { lat: 35.172358507549596, lng: 126.90545394866643 },
-  { lat: 35.15474103200252, lng: 129.11827889154455 },
-  { lat: 36.80711722863776, lng: 127.14020346037576 },
-  { lat: 37.28957415752673, lng: 127.00103752005424 },
-  { lat: 35.83953896766896, lng: 128.7566880321854 },
-  { lat: 37.51027412948879, lng: 127.08227718124704 },
-  { lat: 37.493581783270294, lng: 126.72541955660554 },
-  { lat: 35.135291862962795, lng: 129.10060911448775 },
-  { lat: 35.174574933144065, lng: 126.91389980787773 },
-  { lat: 37.795887691878654, lng: 127.10660416587146 },
-  { lat: 37.59288687521181, lng: 126.96560524627377 },
-  { lat: 37.45076411130452, lng: 127.14593003749792 },
-  { lat: 35.72963747546802, lng: 128.27079056365005 },
-  { lat: 36.02726828142973, lng: 129.37257233594056 },
-  { lat: 35.0708030360945, lng: 129.0593185494088 },
-  { lat: 35.86835862950247, lng: 128.59755089175871 },
-  { lat: 33.51133264696746, lng: 126.51852347452322 },
-  { lat: 37.61284289586752, lng: 127.03120547238589 },
-  { lat: 35.851696038722466, lng: 128.59092937125666 },
-  { lat: 37.59084695083232, lng: 127.01872773588882 },
-  { lat: 37.413461468258156, lng: 126.67735680840826 },
-  { lat: 35.920390371093205, lng: 128.54411720249956 },
-  { lat: 36.65489374054824, lng: 127.48374816871991 },
-  { lat: 37.49491987110441, lng: 127.01493134206048 },
-  { lat: 37.64985695608336, lng: 127.14496345268074 },
-  { lat: 37.55686770317417, lng: 127.16927880543041 },
-  { lat: 37.37014007589146, lng: 127.10614330185591 },
-  { lat: 37.5350236507627, lng: 126.96157681184789 },
-  { lat: 37.40549630594667, lng: 126.8980581820004 },
-  { lat: 34.867950544005744, lng: 128.69069690081176 },
-  { lat: 35.16317059543225, lng: 128.98452978748048 },
-  { lat: 36.607484825953186, lng: 127.48520451195111 },
-  { lat: 35.82667262227336, lng: 127.1030670574823 },
-  { lat: 35.82003554991111, lng: 127.14810974062483 },
-  { lat: 35.097485195649455, lng: 128.99486181862338 },
-  { lat: 37.32204249590605, lng: 127.95591893585816 },
-  { lat: 37.50535127272031, lng: 127.1047465440526 },
-  { lat: 36.99081407156533, lng: 127.09338324956647 },
-  { lat: 37.323486640444834, lng: 127.12285239871076 },
-  { lat: 35.78973089440451, lng: 127.13644319545601 },
-  { lat: 35.641373953578196, lng: 129.35463220719618 },
-  { lat: 37.47423127310911, lng: 126.97625029161996 },
-  { lat: 35.84357192991226, lng: 128.61143720719716 },
-  { lat: 37.180974984085736, lng: 128.20294526341132 },
-  { lat: 37.57895718642583, lng: 126.9316897337244 },
-  { lat: 33.49077253755052, lng: 126.49314817000993 },
-  { lat: 36.42175925330255, lng: 128.67409133225766 },
-  { lat: 37.46405540570109, lng: 126.7153544119173 },
-  { lat: 37.594758776232126, lng: 127.10099917489818 },
-  { lat: 37.7239966558994, lng: 127.0478671731854 },
-  { lat: 35.86680171505329, lng: 128.5923738376741 },
-  { lat: 37.560573727266785, lng: 126.81239107485251 },
-  { lat: 37.78692224857484, lng: 126.98966010341789 },
-  { lat: 35.10368644802913, lng: 129.0206862606022 },
-  { lat: 37.063839948992644, lng: 127.06856523030079 },
-  { lat: 37.34344643728643, lng: 127.94382181350932 },
-  { lat: 37.512521267219064, lng: 127.40054805648133 },
-  { lat: 35.15286653837983, lng: 126.90419903971498 },
-  { lat: 35.173238445546296, lng: 129.176082844468 },
-  { lat: 36.082394201323524, lng: 129.40330471725923 },
-  { lat: 36.09622092762977, lng: 128.43314679004078 },
-  { lat: 37.582855922985544, lng: 126.91907857008522 },
-  { lat: 37.516000983841586, lng: 128.72798872032757 },
-  { lat: 37.48429363675198, lng: 127.0379630203579 },
-  { lat: 37.54502575965604, lng: 126.95429338245707 },
-  { lat: 35.236247173046394, lng: 128.8677618015292 },
-  { lat: 37.40157536691968, lng: 127.11717457214067 },
-  { lat: 36.95191038001258, lng: 127.91064040877527 },
-  { lat: 37.491526492971346, lng: 126.85463749525812 },
-  { lat: 36.124356479753196, lng: 128.09517052346138 },
-  { lat: 37.15715169307048, lng: 128.15853461363773 },
-  { lat: 37.5808156608605, lng: 126.95109705510639 },
-  { lat: 37.46931787249714, lng: 126.89904775044873 },
-  { lat: 35.52195614910054, lng: 129.3209904841746 },
-  { lat: 37.58625703195563, lng: 126.9496035206742 },
-  { lat: 37.28463639199199, lng: 126.85984474757359 },
-  { lat: 35.534169458631226, lng: 129.31169021536095 },
-  { lat: 37.553341234194285, lng: 127.15481222237025 },
-  { lat: 35.180032285898854, lng: 128.06954509175367 },
-];
-
-function getRandomUuid() {
-  return Math.random().toString(36).substring(2, 10);
-}
-
-function getRandomAvailableSpot() {
-  return Math.floor(Math.random() * 50);
-}
-
-export const markerDummyData: MarkerDataType[] = rawPositions.map((pos) => ({
-  parkingLotUuid: getRandomUuid(),
-  availableSpots: getRandomAvailableSpot(),
-  latitude: pos.lat,
-  longitude: pos.lng,
-}));
-
-export const parkingLotSimpleInfoDummyData: ParkingLotModalInfoType = {
-  name: '부산갈매기의 주차장',
-  address: '부산광역시 갈매기 3로 12',
-  distance: 350,
-  thumbnailUrl: 'https://dummyimage.com/90x90',
-  baseFee: 2000,
-  evchargeType: [1, 2],
-  averageRating: 4.5,
-  reviewCount: 12497,
-  likeCount: 1000,
-  dislikeCount: 4,
-};
